Cleanup and style tweak
Former-commit-id: 8d964e871ed5add3b6bff650cca8970464a6afa8

+++ app/Main.hs
@@ -52,7 +52,8 @@ main = runProgram =<< execParser opts
 
 
 runProgram :: ServerOptions -> IO ()
-runProgram (ServerOptions debug False connectionStr port flavour) = startApp connectionStr debug port flavour
+runProgram (ServerOptions debug False connectionStr port flavour) =
+    startApp connectionStr debug (fromIntegral port) flavour
 runProgram _ = migrate
